Moved userclasses routing and controller functions to the myclasses module
