+++ modules/users/server/routes/users.server.routes.js
@@ -7,9 +7,6 @@ module.exports = function (app) {
   // Setting up the users profile api
   app.route('/api/users/me').get(users.me);
   app.route('/api/users').put(users.update);
-  app.route('/api/userclasses').put(users.updateclasses);  // Put for the userclasses
-  app.route('/api/userclasses').get(users.getclasses);     // Grab the user classes
-  app.route('/api/userclasses/deleteclass').put(users.deleteclass); // Delete the user class
   app.route('/api/users/accounts').delete(users.removeOAuthProvider);
   app.route('/api/users/password').post(users.changePassword);
   app.route('/api/users/picture').post(users.changeProfilePicture);
